pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurinspectre"
version = "2.0.0"
authors = [
    {name = "NeurInSpectre Security Team", email = "contact@neurinspectre.dev"},
]
description = "Neural Network Security Analysis Framework - Time Travel Debugger for AI/ML Model Security"
readme = "README.md"
requires-python = ">=3.10,<3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "neural networks",
    "security", 
    "adversarial attacks",
    "machine learning",
    "ai safety",
    "debugging"
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.18.1",
    "torchaudio>=2.3.1",
    "autoattack @ git+https://github.com/fra31/auto-attack",
    "onnxruntime",
    "transformers>=4.30.0",
    "accelerate>=0.29.0",
    "sentence-transformers>=2.2.0",
    "safetensors>=0.3.0",
    "numpy>=1.21.0,<2.0.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "scikit-image>=0.19.0",
    "pandas>=1.3.0",
    "umap-learn>=0.5.3",
    "numba>=0.61.2",
    "llvmlite>=0.44.0",
    "pynndescent>=0.5",
    "sympy>=1.13.0",
    "networkx>=3.0.0",
    "matplotlib>=3.5.0,<3.8.0",
    "plotly>=5.0.0",
    "seaborn>=0.11.0",
    "pillow>=9.0.0",
    "kaleido>=0.2.0",
    "ipywidgets>=7.7.0",
    "opencv-contrib-python-headless==4.8.1.78",
    "dash==2.15.0",
    "dash-core-components==2.0.0",
    "dash-html-components==2.0.0",
    "dash-table==5.0.0",
    "dash-bootstrap-components>=1.4.0",
    "flask<3.1",
    "werkzeug<3.1",
    "click>=8.1.0",
    "rich>=13.7.0",
    "rich-click>=1.8.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "huggingface-hub>=0.12.0",
    "llama-cpp-python==0.3.0",
    "build>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-html>=3.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinxcontrib-napoleon>=0.7",
]
viz = [
    "jupyter>=1.0.0",
    "ipywidgets>=7.7.0",
    "dash-bootstrap-components>=1.4.0",
]

[project.scripts]
neurinspectre = "neurinspectre.cli.main:main"
neurinspectre-obfgrad = "neurinspectre.security.visualization.obfuscated_gradient_visualizer:main"

[project.urls]
Homepage = "https://github.com/packetmaven/NeurInSpectre"
"Bug Reports" = "https://github.com/packetmaven/NeurInSpectre/issues"
Source = "https://github.com/packetmaven/NeurInSpectre"
Documentation = "https://github.com/packetmaven/NeurInSpectre#readme"

[tool.setuptools]
packages = ["neurinspectre"]

[tool.setuptools.package-data]
neurinspectre = ["dataset_cache/*.json", "*.npy"]

[tool.setuptools.data-files]
configs = [
    "configs/red.yaml",
    "configs/blue.yaml",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=neurinspectre",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--verbose"
] 
markers = [
    "slow: slow or long-running tests",
]